lab3/requirements.txt
numpy>=1.19.2
pillow>=8.0.1
matplotlib>=3.3.4
